ID以外のgetterの戻り値をプリミティブ型に変更

+++ src/adapters/gateways/persistance/rdb/model/balance_model.go
@@ -24,7 +24,7 @@ func (Balance) TableName() string {
 func ConvertToBalance(userID uint, remainingAmount uint) Balance {
 	return Balance{
 		UserID: userID,
-		Amount: uint(remainingAmount),
+		Amount: remainingAmount,
 	}
 }
